Move type expression classes to a separate package
