+++ src/main/kotlin/sigma/values/Closure.kt
@@ -2,7 +2,7 @@ package sigma.values
 
 import sigma.Thunk
 import sigma.expressions.Expression
-import sigma.expressions.TupleTypeLiteral
+import sigma.typeExpressions.TupleTypeLiteral
 import sigma.values.tables.Scope
 import sigma.values.tables.Table
 import sigma.values.tables.chainWith
