+++ src/test/kotlin/sigma/expressions/UnorderedTupleTypeLiteralTests.kt
@@ -1,9 +1,9 @@
 package sigma.expressions
 
-import org.junit.jupiter.api.Disabled
 import sigma.StaticTypeScope
-import sigma.TypeExpression
+import sigma.typeExpressions.TypeExpression
 import sigma.TypeReference
+import sigma.typeExpressions.UnorderedTupleTypeLiteral
 import sigma.types.BoolType
 import sigma.types.IntCollectiveType
 import sigma.types.UnorderedTupleType
